refactor: change email infra paths structure
