+++ src/infra/sender-email/adapters/protocols/emailAdapterInterface.ts
